requirements.txt
"""Requirements and dependencies for hardware-aware neural networks."""

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.64.0
pillow>=8.3.0

# Configuration and utilities
pyyaml>=6.0
omegaconf>=2.2.0
pathlib2>=2.3.0

# Model export and deployment
onnx>=1.12.0
onnxruntime>=1.12.0
# tensorflow>=2.10.0  # Optional for TFLite export
# tensorflow-lite>=2.10.0  # Optional for TFLite runtime
# openvino>=2022.1.0  # Optional for OpenVINO
# coremltools>=6.0  # Optional for CoreML export

# Demo and visualization
streamlit>=1.20.0
plotly>=5.10.0

# Development and testing
pytest>=7.0.0
black>=22.0.0
ruff>=0.0.200
pre-commit>=2.20.0

# Optional: Advanced quantization and pruning
# torch-pruning>=1.2.0
# nncf>=2.0.0
# pytorch-quantization>=2.1.0

# Optional: Federated learning
# flower>=1.0.0
# fedml>=0.7.0

# Optional: IoT and messaging
# paho-mqtt>=1.6.0
# kafka-python>=2.0.0

# Optional: Signal processing
# opencv-python>=4.6.0
# scikit-image>=0.19.0
# librosa>=0.9.0
# soundfile>=0.10.0

# Optional: MLOps
# mlflow>=2.0.0
# wandb>=0.13.0
